requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
python-dotenv
Pillow
PyNaCl
bs4
beautifulsoup4
lxml
openai
pydactyl
pteropy
websockets